update app context
