--- freejoas-app/src/utils/UserContext.js
@@ -1,15 +1,0 @@
-import React,{createContext, useContext, useState} from 'react';
-
-const UserContext = createContext();
-
-export const UserProvider = ({children}) => {
-    const [user, setUser] = useState(null);
-
-    return (
-        <UserContext.Provider value={{user, setUser}}>
-            {children}
-        </UserContext.Provider>
-    );
-}
-
-export const useUser = () => useContext(UserContext);
